feat: update comments rather than new
Currently things like new commits or ammending existing commits
results in a new hash and subsequently a new comment. This only
makes it hard to see what is going on in the PR due to many comments.

This should make it that a single comment is made per PR (actually per
workflow too).